NDRS-840: replace hard-coded prococol versions
